Move primitives into lib from test Add factory for semaphore primitive Small clean code
